% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{build_metric_set}
\alias{build_metric_set}
\title{Build a [yarstick::metric_set()] from a list of zero or more charcater function names}
\usage{
build_metric_set(
  x = c("roc_auc", "accuracy", "tss_max"),
  metrics = known_metrics("all")
)
}
\arguments{
\item{x}{chr, one or more quoted names of existing metrics functions}

\item{metrics}{list of known metrics functions, see [known_metrics]}
}
\value{
the output of [yardstick::metric_set()]
}
\description{
This is a hack to allow programmatic selection of common metrics
}
\seealso{
["passing metrics to yarstick"](https://forum.posit.co/t/passing-metrics-into-yardstick/172004)
}
