% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confmat.R
\name{confmat}
\alias{confmat}
\title{Generate confusion matrices}
\usage{
confmat(x, truth, estimate)
}
\arguments{
\item{x}{fitted_workflowset}

\item{truth, estimate}{see [yardstick::conf_mat()] for details, not required if `x` is a last_fit_set}
}
\value{
tibble with three elements: wflow_id, conf_mat, plot
}
\description{
Borrows heavily from the [yardstick package](https://github.com/tidymodels/yardstick/blob/51761c4e7a34e960949c75aeb2952ef02c408106/R/conf_mat.R#L421)
}
