% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confmat.R
\name{confmat_as_tibble}
\alias{confmat_as_tibble}
\title{Convert a confusion matrix (conf_mat class) to a tibble}
\usage{
confmat_as_tibble(cm)
}
\arguments{
\item{cm}{conf_mat class object}
}
\value{
data frame (tibble) of the confusion matrix
}
\description{
Convert a confusion matrix (conf_mat class) to a tibble
}
