% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_workflowsets.R
\name{extract_predictions}
\alias{extract_predictions}
\title{Extract predictions}
\usage{
extract_predictions(
  x,
  truth = "patch",
  estimate = ".pred_class",
  outnames = c("class", ".pred_class")
)
}
\arguments{
\item{x}{fitted workflow or fitted_workflowset data frame}

\item{truth}{chr the name of the truth column (usually a factor)}

\item{estimate}{chr, the predicted probablility column name}

\item{outnames}{chr, names the output data frame should have}
}
\value{
data frame
}
\description{
Extract predictions
}
