% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_type.R
\name{get_model_type}
\alias{get_model_type}
\alias{get_model_type.default}
\alias{get_model_type.workflow}
\alias{get_model_type.workflow_set}
\title{A convenience to quickly get the type name of a model}
\usage{
get_model_type(x)

\method{get_model_type}{default}(x)

\method{get_model_type}{workflow}(x)

\method{get_model_type}{workflow_set}(x)
}
\arguments{
\item{x}{object (model, workflow, workflowset)}
}
\value{
string value of the model type
}
\description{
A convenience to quickly get the type name of a model
}
