% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{known_metrics}
\alias{known_metrics}
\title{Retrieve a list of known metrics}
\usage{
known_metrics(x = c("all", "roc_auc", "accuracy", "tss_max")[1])
}
\arguments{
\item{x}{chr, one or more quoted names of existing metrics functions.  Use "all" (the default)
to retrieve them all}
}
\value{
a named list of metrics functions
}
\description{
The default set includes [yardstick::roc_auc()], [yardstick::accuracy()] and [tidysdm::tss_max()].
Others may be added by rebuilding the package.
}
