% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muddler-package.R
\docType{package}
\name{muddler}
\alias{muddler-package}
\alias{muddler}
\title{muddler package}
\description{
Helpful functions for working with tidymodels.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/muddler}
}

}
\author{
\strong{Maintainer}: Tupper Ben \email{btupper@bigelow.org}

}
