% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{read_workflow}
\alias{read_workflow}
\alias{write_workflow}
\title{Read or write a workflow given the version and path}
\usage{
read_workflow(filename)

write_workflow(x, filename)
}
\arguments{
\item{filename}{str the path to the workflow}

\item{x}{workflow object}
}
\value{
workflow
}
\description{
Read or write a workflow given the version and path
}
