% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocauc.R
\name{rocauc}
\alias{rocauc}
\title{Generate AUC of ROC}
\usage{
rocauc(x, truth = "patch", prob = ".pred_1")
}
\arguments{
\item{x}{fitted_workflowset or resample_results}

\item{truth, prob}{see [yardstick::roc_auc()] for details, not required if `x` is a last_fit_set}
}
\value{
tibble with four elements: wflow_id, auc, roc, plot
}
\description{
Generate AUC of ROC
}
