% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowset.R
\name{workflowset_selectomatic}
\alias{workflowset_selectomatic}
\title{Selects best hyperparams, transfers those to the final model and then 
performs a final fit and testing suite for each wflow in the set.}
\usage{
workflowset_selectomatic(wflow, splitdata)
}
\arguments{
\item{wflow}{a workflow set with one or more workflows}

\item{splitdata}{rsplit object with training and testing data}
}
\value{
a data frame with one or more  "last_fit" model objects
}
\description{
Note that this wraps a number of important steps into one.  We provided this
in the interest of speed, but the details are worthy of investigating.
}
