% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_workflowsets.R
\name{write_fitted_workflowset}
\alias{write_fitted_workflowset}
\alias{read_last_fit_set}
\title{Read and write a `fitted_workflowset` table}
\usage{
write_fitted_workflowset(x, filename = "fitted_workflowset.rds")

read_last_fit_set(filename = "fitted_workflowset.rds")
}
\arguments{
\item{x}{a last_fit_set date frame or one or more last_fit sets}

\item{filename}{chr, the name of the file to write to (default is "last_fit_set.rds")}
}
\value{
last_fit_set date frame
}
\description{
Read and write a `fitted_workflowset` table

Write a `fitted_workflowset` table
}
