% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confmat.R
\name{confmat}
\alias{confmat}
\title{Given a table of model fits, plot the confusion matrices}
\usage{
confmat(x, truth, estimate, ...)
}
\arguments{
\item{x}{fitted workflow or fitted workflow_set}

\item{truth, estimate}{see [yardstick::conf_mat()] for details}
}
\value{
tibble with three elements: workflow, confmat, plot
}
\description{
Borrows heavily from the [yardstick package](https://github.com/tidymodels/yardstick/blob/51761c4e7a34e960949c75aeb2952ef02c408106/R/conf_mat.R#L421)
}
